requirements.txt
Flask==2.3.3
line-bot-sdk==3.5.0
opencv-python==4.8.1.78
numpy>=1.26.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
insightface==0.7.3
onnxruntime>=1.17.0
tqdm==4.66.1
Pillow==10.0.1
scikit-learn>=1.3.2